FINAL refactoring - added comments and optimizations.
