+++ Battleship.model/grid/SquareCoordinate.cs
@@ -1,0 +1,4 @@
+﻿namespace Battleship.Model;
+
+/// <summary>Represents a grid coordinate identified by row and column.</summary>
+public record SquareCoordinate(int Row, int Column);
