+++ Battleship.sln
@@ -1,13 +1,51 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.5.33502.453
+VisualStudioVersion = 17.14.37216.2
 MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Battleship.model", "Battleship.model.csproj", "{5135EB26-709F-8799-0F74-90AF715D2654}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Battleship.Tests", "Battleship.Tests\Battleship.Tests.csproj", "{BE8CD0D9-C5E6-4EFA-808E-E8168CC046CA}"
+EndProject
 Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{5135EB26-709F-8799-0F74-90AF715D2654}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5135EB26-709F-8799-0F74-90AF715D2654}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5135EB26-709F-8799-0F74-90AF715D2654}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{5135EB26-709F-8799-0F74-90AF715D2654}.Debug|x64.Build.0 = Debug|Any CPU
+		{5135EB26-709F-8799-0F74-90AF715D2654}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{5135EB26-709F-8799-0F74-90AF715D2654}.Debug|x86.Build.0 = Debug|Any CPU
+		{5135EB26-709F-8799-0F74-90AF715D2654}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5135EB26-709F-8799-0F74-90AF715D2654}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5135EB26-709F-8799-0F74-90AF715D2654}.Release|x64.ActiveCfg = Release|Any CPU
+		{5135EB26-709F-8799-0F74-90AF715D2654}.Release|x64.Build.0 = Release|Any CPU
+		{5135EB26-709F-8799-0F74-90AF715D2654}.Release|x86.ActiveCfg = Release|Any CPU
+		{5135EB26-709F-8799-0F74-90AF715D2654}.Release|x86.Build.0 = Release|Any CPU
+		{BE8CD0D9-C5E6-4EFA-808E-E8168CC046CA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BE8CD0D9-C5E6-4EFA-808E-E8168CC046CA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BE8CD0D9-C5E6-4EFA-808E-E8168CC046CA}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{BE8CD0D9-C5E6-4EFA-808E-E8168CC046CA}.Debug|x64.Build.0 = Debug|Any CPU
+		{BE8CD0D9-C5E6-4EFA-808E-E8168CC046CA}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{BE8CD0D9-C5E6-4EFA-808E-E8168CC046CA}.Debug|x86.Build.0 = Debug|Any CPU
+		{BE8CD0D9-C5E6-4EFA-808E-E8168CC046CA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BE8CD0D9-C5E6-4EFA-808E-E8168CC046CA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BE8CD0D9-C5E6-4EFA-808E-E8168CC046CA}.Release|x64.ActiveCfg = Release|Any CPU
+		{BE8CD0D9-C5E6-4EFA-808E-E8168CC046CA}.Release|x64.Build.0 = Release|Any CPU
+		{BE8CD0D9-C5E6-4EFA-808E-E8168CC046CA}.Release|x86.ActiveCfg = Release|Any CPU
+		{BE8CD0D9-C5E6-4EFA-808E-E8168CC046CA}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {52B90A9C-75CC-4915-8271-0A0ED93075F2}
+		SolutionGuid = {56E9969A-216F-410D-A8B1-551606D89426}
 	EndGlobalSection
 EndGlobal
