+++ Battleship.tests/grid/SquareTests.cs
@@ -1,0 +1,19 @@
+using Battleship.Model;
+
+namespace Battleship.Tests;
+
+[TestClass]
+public sealed class SquareTests
+{
+    [TestMethod]
+    public void ConstructorCreateSquareWithRowAndColumnProvided()
+    {
+        int row = 4;
+        int column = 8;
+
+        var square = new Square(row, column);
+
+        Assert.AreEqual(row, square.Row);
+        Assert.AreEqual(column, square.Column);
+    }
+}
