--- model/ITargetSelector.cs
@@ -1,8 +1,0 @@
-﻿using model;
-
-namespace model;
-
-public interface ITargetSelector
-{
-    Square Next();
-}
