--- model/InlineTargetSelector.cs
@@ -1,11 +1,0 @@
-﻿using model;
-
-namespace model;
-
-public class InlineTargetSelector : ITargetSelector
-{
-    public Square Next()
-    {
-        throw new NotImplementedException();
-    }
-}
