--- model/RandomTargetSelector.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using model;
-
-namespace model;
-
-public class RandomTargetSelector : ITargetSelector
-{
-    public RandomTargetSelector(ShotsGrid grid, int shipLength)
-    {
-        this.grid = grid;
-        this.shipLength = shipLength;
-    }
-    public Square Next()
-    {
-        var placements = grid.GetAvailablePlacements(shipLength);
-        var candidates = placements.SelectMany(s => s);
-        var selectedIndex = random.Next(candidates.Count());
-        return candidates.ElementAt(selectedIndex);
-    }
-    private readonly ShotsGrid grid;
-    private readonly int shipLength;
-    private readonly Random random = new Random();
-}
