--- model/SquareCoordinate.cs
@@ -1,3 +1,0 @@
-﻿namespace model;
-
-public record SquareCoordinate(int Row, int Column);
