--- tests/MSTestSetting.cs
@@ -1,1 +1,0 @@
-[assembly: Parallelize(Scope = ExecutionScope.MethodLevel)]
